Remove compat.mjs
Replaces the unique() function (which should either be an abstract
operation, or open-coded.) There's nothing "compat" about it anyway as
there exists no Array.unique proposal that I am aware of.
