--- polyfill/lib/compat.mjs
@@ -1,12 +1,0 @@
-export function unique(arr) {
-  const obj = Object.create(null);
-  for (var i = 0; i < arr.length; i++) {
-    var v = arr[i];
-    obj[v] = v;
-  }
-  var res = [];
-  for (var p in obj) {
-    res.push(obj[p]);
-  }
-  return res;
-}
